+++ examples/logical_interconnects.go
@@ -32,7 +32,7 @@ func main() {
 		os.Getenv("ONEVIEW_OV_DOMAIN"),
 		os.Getenv("ONEVIEW_OV_ENDPOINT"),
 		false,
-		1000,
+		1200,
 		"*")
 
 	fmt.Println("....  Logical Interconnects Collection .....")