+++ CHANGELOG.md
@@ -2,6 +2,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 # [v1.2.0] (Unreleased)
+### Notes
+- This version partially supports API1200.
 
 ### Oneview Features supported
 - Enclosure