+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Oneview Features supported
 - Enclosure
+- FC Network
+- FCoE Network
+- Interconnect type
 - Network Set
 - Server Profile
 - Task