+++ CHANGELOG.md
@@ -7,9 +7,14 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Oneview Features supported
 - Enclosure
+- Interconnect
+- Logical interconnect
 - Network Set
+- Server hardware
+- Server hardware type
 - Server Profile
 - Task
+- Uplink set
 
 # [v1.1.0] (03-12-2019)
 ### Notes